Add check for connectivity
The gist of the checking: agents (i.e. pods) sends their names in
requests payload; those name is unique identifier for the agent. Then
the responses from agents compared to data obtained from k8s api. In
order the check to be successful two criteria must be met: response from
particular pod must be present and it must not be older than interval of
heart beating for agents - 1 minute.

Change-Id: I38bb872e508f97346a2b820635a88378d3bba23d